+++ cfgov/v1/models/browse_filterable_page.py
@@ -12,7 +12,6 @@ from wagtail.wagtailadmin.edit_handlers import TabbedInterface, ObjectList
 from . import base, molecules, organisms, ref
 from .learn_page import AbstractFilterPage
 from .. import forms
-from ..util.util import get_secondary_nav_items
 from ..util import filterable_context