+++ cfgov/v1/models/sublanding_filterable_page.py
@@ -14,7 +14,6 @@ from wagtail.wagtailimages.blocks import ImageChooserBlock
 from . import base, molecules, organisms, ref
 from .learn_page import AbstractFilterPage
 from .. import forms
-from ..util.util import get_secondary_nav_items
 from ..util import filterable_context
 
 from .base import CFGOVPage